+++ pica-record/fuzz/fuzz_targets/fuzz_field_ref.rs
@@ -1,7 +1,7 @@
 #![no_main]
 use libfuzzer_sys::fuzz_target;
-use pica_record::TagRef;
+use pica_record::FieldRef;
 
 fuzz_target!(|data: &[u8]| {
-    let _tag = TagRef::from_bytes(data);
+    let _field = FieldRef::from_bytes(data);
 });