+++ pica-record/CHANGELOG.md
@@ -6,5 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `Field` and `FieldRef`
+* Add `OccurenceRef` and `OccurenceRef`
 * Add `SubfieldRef` and `Subfield`
 * Add `TagRef` and `Tag`